[#3] Refactor: onClickHandler 생성하는 useAddProduct 훅 추가
