+++ src/components/product/ProductItem.jsx
@@ -1,34 +1,27 @@
+import useAddProduct from '../../hooks/useAddProduct';
 import ProductName from '../common/ProductName';
 import ProductNumber from '../common/ProductNumber';
 import ProductPrice from '../common/ProductPrice';
 import ThumbNailImage from '../common/ThumbNailImage';
-import { useProductActionContext } from './../../contexts/ProductContext';
-import LazyImage from './../common/LazyImage';
 import ProductDetail from './ProductDetail';
 import {
   Card,
   CardBody,
   CardFooter,
   Stack,
-  Text,
   IconButton,
   Tag,
   useDisclosure,
-  AspectRatio,
 } from '@chakra-ui/react';
 import React from 'react';
 import { RiShoppingBag2Fill } from 'react-icons/ri';
 
 const ProductItem = ({ product }) => {
-  const { idx, name, mainImage, price, spaceCategory } = product;
+  const { idx, name, price, spaceCategory } = product;
 
   const { isOpen, onOpen, onClose } = useDisclosure();
-  const { addProduct } = useProductActionContext();
 
-  const onClickHandler = e => {
-    e.stopPropagation();
-    addProduct(product);
-  };
+  const onClickHandler = useAddProduct(product);
 
   return (
     <>
