+++ src/hooks/useAddProduct.js
@@ -1,0 +1,14 @@
+import { useProductActionContext } from '../contexts/ProductContext';
+
+const useAddProduct = product => {
+  const { addProduct } = useProductActionContext();
+
+  const onClickHandler = e => {
+    e.stopPropagation();
+    addProduct(product);
+  };
+
+  return onClickHandler;
+};
+
+export default useAddProduct;
